pyproject.toml
[project]
name = "poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "pydantic>=2.11.3",
    "pydantic-ai>=0.1.3",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "reportlab>=4.4.0",
    "streamlit>=1.44.1",
    "torch>=2.6.0",
]
